requirements.txt
certifi==2020.12.5
chardet==4.0.0
facenet-pytorch==2.5.2
idna==2.10
joblib==1.0.1
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.1
threadpoolctl==2.1.0
torch==1.8.0
torchvision==0.9.0
typing-extensions==3.7.4.3
urllib3==1.26.4
